Rename BeetStore to WalletStore
