+++ src/store/index.js
@@ -1,13 +1,13 @@
 import Vue from 'vue/dist/vue.js';
 import Vuex from 'vuex';
-import BeetStore from './modules/BeetStore.js';
+import WalletStore from './modules/WalletStore.js';
 import SettingsStore from './modules/SettingsStore.js';
 
 Vue.use(Vuex);
 
 const store = new Vuex.Store({
     modules: {
-        BeetStore,
+        WalletStore,
         SettingsStore
     }
   });
